chore: pass test result to docs
